+++ src/ui/components/shared/SharingModal/CopyUrl.tsx
@@ -1,0 +1,36 @@
+import React, { useState } from "react";
+
+export default function CopyUrl({ recordingId }: { recordingId: string }) {
+  const [copyClicked, setCopyClicked] = useState(false);
+
+  const handleCopyClick = () => {
+    navigator.clipboard.writeText(`https://replay.io/view?id=${recordingId}`);
+    setCopyClicked(true);
+    setTimeout(() => setCopyClicked(false), 2000);
+  };
+
+  if (copyClicked) {
+    return (
+      <div className="copy-link">
+        <div className="status">
+          <div className="success">Link copied</div>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="copy-link">
+      <input
+        type="text"
+        value={`https://replay.io/view?id=${recordingId}`}
+        onKeyPress={e => e.preventDefault()}
+        onChange={e => e.preventDefault()}
+      />
+      <button onClick={handleCopyClick}>
+        <div className="img link" />
+        <div className="label">Copy Link</div>
+      </button>
+    </div>
+  );
+}
