+++ src/ui/components/shared/SharingModal/RecordingPrivacy.css
@@ -1,0 +1,71 @@
+.privacy-toggle-container {
+  border-bottom: 1px solid var(--theme-splitter-color);
+  padding-bottom: 20px;
+}
+
+.privacy-toggle {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  margin-bottom: 12px;
+}
+
+.privacy-toggle > :not(:first-child) {
+  margin-left: 8px;
+}
+
+.privacy-toggle .label {
+  flex-grow: 1;
+}
+
+.modal-container .privacy-toggle button.toggle {
+  position: relative;
+  height: 18px;
+  border-radius: 12px;
+  width: 28px;
+  align-items: center;
+  transition-duration: 200ms;
+  padding: 0px;
+  box-sizing: content-box;
+}
+
+.privacy-note.is-public,
+.privacy-note.is-public .label-description,
+.privacy-note.is-public .label-description a:visited {
+  color: #7d000f;
+}
+
+.privacy-note {
+  margin-left: 24px;
+}
+
+.privacy-toggle .toggle:focus {
+  outline: none;
+}
+
+.privacy-toggle .toggle.on {
+  border: 2px solid var(--dark-blue);
+  background: var(--dark-blue);
+}
+
+.privacy-toggle .toggle.off {
+  border: 2px solid var(--light-grey);
+  background: var(--light-grey);
+}
+
+.privacy-toggle .toggle.on .switch {
+  left: 10px;
+}
+
+.privacy-toggle .toggle.off .switch {
+  left: 0px;
+}
+
+.privacy-toggle  .switch {
+  position: relative;
+  height: 18px;
+  width: 18px;
+  border-radius: 50%;
+  background: white;
+  transition-duration: 200ms;
+}
