+++ src/ui/components/shared/SharingModal/index.tsx
@@ -1,0 +1,3 @@
+import SharingModal from "./SharingModal";
+
+export default SharingModal;
