+++ src/ui/state/session.ts
@@ -1,0 +1,5 @@
+export interface User {
+  name: string;
+  picture: string;
+  id: string;
+}
